requirements.txt
python-dotenv
streamlit
langchain_google_genai
google.generativeai
pandas